\name{getDependentVariables}
\alias{getDependentVariables}
\title{Get Dependent Variables from a Zelig-style Formula...}
\usage{getDependentVariables(x, ...)}
\description{Get Dependent Variables from a Zelig-style Formula}
\details{This method...}
\value{a character-vector}
\arguments{\item{a}{list or formula}}
