\name{zelig2aov}
\alias{zelig2aov}
\title{Interface between the Zelig Model aov and...}
\usage{zelig2aov(formula, ..., data)}
\description{Interface between the Zelig Model aov and 
the Pre-existing Model-fitting Method}
\value{a list specifying '.function'}
\arguments{\item{formula}{a formula}
\item{...}{additonal parameters}
\item{data}{a data.frame}}
