\name{zelig2sur}
\alias{zelig2sur}
\title{Interface between the Zelig Model sur and...}
\usage{zelig2sur(formula, ..., data)}
\description{Interface between the Zelig Model sur and 
the Pre-existing Model-fitting Method}
\value{a list specifying '.function'}
\arguments{\item{formula}{a formula}
\item{...}{additonal parameters}
\item{data}{a data.frame}}
