\name{zelig2threesls}
\alias{zelig2threesls}
\title{Interface between the Zelig Model threesls and...}
\usage{zelig2threesls(formula, ..., data)}
\description{Interface between the Zelig Model threesls and 
the Pre-existing Model-fitting Method}
\value{a list specifying '.function'}
\arguments{\item{formula}{a formula}
\item{...}{additonal parameters}
\item{data}{a data.frame}}
