\name{zelig2twosls}
\alias{zelig2twosls}
\title{Interface between the Zelig Model twosls and...}
\usage{zelig2twosls(formula, ..., data)}
\description{Interface between the Zelig Model twosls and 
the Pre-existing Model-fitting Method}
\value{a list specifying '.function'}
\arguments{\item{formula}{a formula}
\item{...}{additonal parameters}
\item{data}{a data.frame}}
