\name{ZeligLeastSquares-package}
\title{ZeligLeastSquares...}
\description{ZeligLeastSquares}
\details{\tabular{ll}{
Package: \tab ZeligLeastSquares\cr
Version: \tab 0.1\cr
Date: \tab 2011-04-25\cr
Depends: Zelig
License: \tab GPL version 2 or newer\cr
}

Edit this description}
\alias{ZeligLeastSquares-package}
\alias{ZeligLeastSquares}
\docType{package}
\author{UNKNOWN AUTHOR}
\keyword{package}
